rename : ui component file
