--- packages/ui/src/components/button sample/index.ts
@@ -1,1 +1,0 @@
-export { default as ButtonSample } from "./ButtonSample";
