+++ packages/ui/src/components/button/Button.stories.tsx
@@ -1,7 +1,7 @@
 import type { Meta, StoryObj } from "@storybook/react";
 import { FiUserPlus } from "react-icons/fi";
 import { story } from "../../styles/storyStyle.css";
-import Button from "./Button";
+import Button from "./button";
 
 const meta: Meta<typeof Button> = {
   component: Button,
