+++ packages/ui/src/components/button/index.ts
@@ -1,1 +1,1 @@
-export { default as Button } from "./Button";
+export { default as Button } from "./button";
