+++ packages/ui/src/components/checkBox/index.ts
@@ -1,1 +1,1 @@
-export { default as CheckBox } from "./CheckBox";
+export { default as CheckBox } from "./checkbox";
