+++ packages/ui/src/components/dialog/Dialog.stories.tsx
@@ -1,7 +1,7 @@
 import type { Meta, StoryObj } from "@storybook/react";
 import { useState } from "react";
 import { Button } from "../button";
-import Dialog from "./Dialog";
+import Dialog from "./dialog";
 
 const meta: Meta<typeof Dialog> = {
   component: Dialog,
