+++ packages/ui/src/components/dialog/dialog-component/index.ts
@@ -1,0 +1,4 @@
+export { default as DialogBody } from "./dialog-body";
+export { default as DialogFooter } from "./dialog-footer";
+export { default as DialogHeader } from "./dialog-header";
+export { default as DialogWrapper } from "./dialog-wrapper";
