--- packages/ui/src/components/dialog/dialogComponent/index.ts
@@ -1,4 +1,0 @@
-export { default as DialogBody } from "./DialogBody";
-export { default as DialogFooter } from "./DialogFooter";
-export { default as DialogHeader } from "./DialogHeader";
-export { default as DialogWrapper } from "./DialogWrapper";
