+++ packages/ui/src/components/dialog/index.ts
@@ -1,2 +1,2 @@
-export { default as Dialog } from "./Dialog";
-export * from "./dialogComponent";
+export { default as Dialog } from "./dialog";
+export * from "./dialog-component";
