+++ packages/ui/src/components/file-upload/index.ts
@@ -1,0 +1,1 @@
+export { default as FileUpload } from "./file-upload";
