--- packages/ui/src/components/fileUpload/index.ts
@@ -1,1 +1,0 @@
-export { default as FileUpload } from "./FileUpload";
