+++ packages/ui/src/components/index.ts
@@ -1,8 +1,7 @@
 export * from "./button";
-export * from "./button sample";
-export * from "./checkBox";
+export * from "./checkbox";
 export * from "./dialog";
-export * from "./fileUpload";
+export * from "./file-upload";
 export * from "./portal";
 export * from "./radio";
 export * from "./select";
