+++ packages/ui/src/components/portal/index.ts
@@ -1,1 +1,1 @@
-export { default as Portal } from "./Portal";
+export { default as Portal } from "./portal";
