+++ packages/ui/src/components/radio/index.ts
@@ -1,1 +1,1 @@
-export { default as Radio } from "./Radio";
+export { default as Radio } from "./radio";
