+++ packages/ui/src/components/select/Select.stories.tsx
@@ -1,7 +1,7 @@
 import type { Meta, StoryObj } from "@storybook/react";
 import { useState } from "react";
 import { Color } from "../../styles";
-import Select from "./Select";
+import Select from "./select";
 
 const meta: Meta<typeof Select> = {
   component: Select,
