+++ packages/ui/src/components/select/index.ts
@@ -1,1 +1,1 @@
-export { default as Select } from "./Select";
+export { default as Select } from "./select";
