+++ packages/ui/src/components/switch/index.ts
@@ -1,1 +1,1 @@
-export { default as Switch } from "./Switch";
+export { default as Switch } from "./switch";
