+++ packages/ui/src/components/tag/index.ts
@@ -1,1 +1,1 @@
-export { default as Tag } from "./Tag";
+export { default as Tag } from "./tag";
