+++ packages/ui/src/components/textfield/index.ts
@@ -1,2 +1,2 @@
-export * from "./textInput";
+export * from "./text-input";
 export * from "./textarea";
