+++ packages/ui/src/components/textfield/text-input/index.ts
@@ -1,0 +1,1 @@
+export { default as TextInput } from "./text-input";
