--- packages/ui/src/components/textfield/textInput/index.ts
@@ -1,1 +1,0 @@
-export { default as TextInput } from "./TextInput";
