+++ packages/ui/src/components/textfield/textarea/Textarea.stories.tsx
@@ -1,6 +1,6 @@
 import type { Meta, StoryObj } from "@storybook/react";
 import { Color } from "../../../styles";
-import Textarea from "./Textarea";
+import Textarea from "./text-area";
 
 const meta: Meta<typeof Textarea> = {
   component: Textarea,
