+++ packages/ui/src/components/textfield/textarea/index.ts
@@ -1,1 +1,1 @@
-export { default as Textarea } from "./Textarea";
+export { default as Textarea } from "./text-area";
