+++ packages/ui/src/components/toast/Toast.stories.tsx
@@ -1,6 +1,6 @@
 import type { Meta, StoryObj } from "@storybook/react";
 import { story } from "../../styles/storyStyle.css";
-import Toast from "./Toast";
+import Toast from "./toast";
 
 const meta: Meta<typeof Toast> = {
   component: Toast,
