+++ packages/ui/src/components/toast/index.ts
@@ -1,1 +1,1 @@
-export { default as Toast } from "./Toast";
+export { default as Toast } from "./toast";
